requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
boto3==1.16.21
botocore==1.19.21
certifi==2020.12.5
click==7.1.2
dj-database-url==0.5.0
Django==3.1.5
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.6.0
django-filter==2.4.0
django-jazzmin==2.4.3
django-nested-admin==3.3.2
django-ordered-model==3.4.1
django-rest-framework==0.1.0
django-storages==1.10.1
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
flake8==3.8.4
gunicorn==20.0.4
jmespath==0.10.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.3.1
python-monkey-business==1.0.0
pytz==2020.5
regex==2020.11.13
s3transfer==0.3.3
sendgrid==6.4.8
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
